refactor: repait ui and use latest data-fair lib packages
