+++ api/doc/agg-result/index.ts
