+++ api/doc/index.ts
@@ -1,12 +1,2 @@
 export type { AggQuery } from './agg-query/index.js'
 export type { AggResult } from './agg-result/index.js'
-
-import * as aggQuery from './agg-query/index.js'
-import * as aggResult from './agg-result/index.js'
-import * as listResponse from './list-response/index.js'
-
-export default {
-  aggQuery,
-  aggResult,
-  listResponse
-}
