--- ui-old/Dockerfile
@@ -1,36 +1,0 @@
-##########################
-# install root dev deps and build
-FROM node:22.9.0-alpine3.19 AS builder
-
-RUN npm i -g clean-modules@3.0.4
-
-WORKDIR /app/
-ADD /package.json .
-ADD /package-lock.json .
-ADD /ui/package.json ui/
-
-RUN npm -w ui --include-workspace-root ci --no-audit --no-fund
-ADD /ui/ ui/
-RUN npm -w ui run build
-
-##########################
-# copy built app and assemble actual dist
-FROM node:22.9.0-alpine3.19
-
-ENV NODE_ENV production
-ENV PORT 8080
-
-WORKDIR /app/
-
-COPY --from=builder /app/ui/.output /app/ui/.output
-
-ADD README.md LICENSE package.json BUILD.json* ./
-
-WORKDIR /app/ui
-
-EXPOSE 8080
-
-USER node
-
-CMD ["node", "--max-http-header-size", "64000", ".output/server/index.mjs"]
-
