--- ui-old/assets/global.scss
@@ -1,1 +1,0 @@
-// Optional global scss settings
