--- ui-old/assets/safe-decode-uri-component.js
@@ -1,7 +1,0 @@
-export default (text) => {
-  try {
-    return decodeURIComponent(text)
-  } catch (err) {
-    return text
-  }
-}
