--- ui-old/assets/settings.scss
@@ -1,5 +1,0 @@
-// Specific Vuetify scss settings
-@forward 'vuetify/settings' with (
-    $color-pack: false,
-    $body-font-family: 'Nunito'
-);
