--- ui-old/package.json
@@ -1,35 +1,0 @@
-{
-  "type": "module",
-  "scripts": {
-    "build": "nuxt build",
-    "dev": "nuxt dev --port 6220",
-    "generate": "nuxt generate",
-    "preview": "nuxt preview",
-    "postinstall": "nuxt prepare",
-    "lint": "eslint .",
-    "lint-fix": "eslint --fix ."
-  },
-  "devDependencies": {
-    "@mdi/font": "^7.4.47",
-    "@nuxtjs/google-fonts": "^3.1.3",
-    "@nuxtjs/i18n": "^8.0.0-rc.11",
-    "@types/debug": "^4.1.7",
-    "@types/truncate-middle": "^1.0.1",
-    "chart.js": "^4.4.1",
-    "dayjs": "^1.11.10",
-    "debug": "^4.3.4",
-    "eslint-plugin-vuetify": "github:albanm/eslint-plugin-vuetify",
-    "iframe-resizer": "^4.3.9",
-    "jwt-decode": "^4.0.0",
-    "nuxt": "^3.8.2",
-    "sass": "^1.69.5",
-    "truncate-middle": "^1.0.6",
-    "upath": "^2.0.1",
-    "vite-plugin-vuetify": "^2.0.1",
-    "vue": "^3.4.5",
-    "vuetify": "^3.4.9"
-  },
-  "dependencies": {
-    "universal-cookie": "^7.2.0"
-  }
-}
