--- ui-old/plugins/chartjs.js
@@ -1,5 +1,0 @@
-import { Chart, BarController, CategoryScale, LinearScale, BarElement, Legend, Tooltip } from 'chart.js'
-
-export default defineNuxtPlugin(() => {
-  Chart.register(BarController, CategoryScale, LinearScale, BarElement, Legend, Tooltip)
-})
