--- ui-old/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  // https://nuxt.com/docs/guide/concepts/typescript
-  "extends": "./.nuxt/tsconfig.json",
-  "compilerOptions": {
-    "allowJs": true,
-    "checkJs": true,
-    "moduleResolution": "bundler",
-    "paths": {
-      "#api/types": ["../api/types/index.ts"],
-      "#api/doc": ["../api/doc/index.ts"]
-    }
-  }
-}
