+++ ui/dts/auto-imports.d.ts
@@ -24,6 +24,9 @@ declare global {
   const dailyApiMetrics: typeof import('../src/components/daily-api-metrics.vue')['default']
   const defineAsyncComponent: typeof import('vue')['defineAsyncComponent']
   const defineComponent: typeof import('vue')['defineComponent']
+  const dfPersonalMenu: typeof import('@data-fair/lib-vuetify/personal-menu.vue')['default']
+  const dfTutorialAlert: typeof import('@data-fair/lib-vuetify/tutorial-alert.vue')['default']
+  const dfUserAvatar: typeof import('@data-fair/lib-vuetify/user-avatar.vue')['default']
   const effectScope: typeof import('vue')['effectScope']
   const filterDatePicker: typeof import('../src/components/filter/filter-date-picker.vue')['default']
   const filterPeriod: typeof import('../src/components/filter/filter-period.vue')['default']
@@ -74,6 +77,7 @@ declare global {
   const triggerRef: typeof import('vue')['triggerRef']
   const unref: typeof import('vue')['unref']
   const useAttrs: typeof import('vue')['useAttrs']
+  const useConceptFilters: typeof import('@data-fair/lib-vue/concept-filters.js')['useConceptFilters']
   const useCssModule: typeof import('vue')['useCssModule']
   const useCssVars: typeof import('vue')['useCssVars']
   const useI18n: typeof import('vue-i18n')['useI18n']
@@ -87,6 +91,7 @@ declare global {
   const useSession: typeof import('@data-fair/lib-vue/session.js')['useSession']
   const useSlots: typeof import('vue')['useSlots']
   const useTemplateRef: typeof import('vue')['useTemplateRef']
+  const useUiNotif: typeof import('@data-fair/lib-vue/ui-notif.js')['useUiNotif']
   const watch: typeof import('vue')['watch']
   const watchEffect: typeof import('vue')['watchEffect']
   const watchPostEffect: typeof import('vue')['watchPostEffect']
