+++ ui/package.json
@@ -12,8 +12,8 @@
     "lint-fix": "eslint --fix ."
   },
   "dependencies": {
-    "@data-fair/lib-vue": "^1.0.1",
-    "@data-fair/lib-vuetify": "^1.0.1",
+    "@data-fair/lib-vue": "^1.1.1",
+    "@data-fair/lib-vuetify": "^1.1.0",
     "@intlify/unplugin-vue-i18n": "^5.2.0",
     "@mdi/js": "^7.4.47",
     "@types/config": "^3.3.5",
