+++ ui/src/App.vue
@@ -2,6 +2,11 @@
   <v-app>
     <v-main>
       <RouterView />
+      <ui-notif />
     </v-main>
   </v-app>
 </template>
+
+<script setup lang="ts">
+import uiNotif from '@data-fair/lib-vuetify/ui-notif.vue'
+</script>
