--- ui/src/main.scss
@@ -1,10 +1,0 @@
-@use 'vuetify' with (
-  $color-pack: false,
-  $body-font-family: 'Nunito'
-);
-
-// v-icon has an opacity when it is prepended in a list of items.
-// ref: https://github.com/vuetifyjs/vuetify/issues/18760
-.v-icon {
-  --v-medium-emphasis-opacity: 1 !important;
-}
